# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Arbor API
  summary: 多租户人设工作台 HTTP 契约
  version: 0.1.0
  description: |
    入站适配器契约。实现必须把租户与人设边界留在应用层/仓储端口，
    而不是仅在文档里声明。
servers:
  - url: /v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    TenantHeader:
      name: X-Tenant-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    PersonaId:
      name: persona_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string
              description: 每次错误单独生成的 ULID
...
    PersonaSkin:
      type: string
      enum: [companion, employee]
    Persona:
      type: object
      properties:
        id:
          type: string
          format: uuid
        tenant_id:
          type: string
          format: uuid
        skin:
          $ref: "#/components/schemas/PersonaSkin"
        display_name:
          type: string
        one_liner:
          type: string
        personality:
          type: object
          additionalProperties: true
        taboos:
          type: array
          items:
            type: string
        relationships:
          type: array
          items:
            type: object
            additionalProperties: true
        avatar:
          type: string
          description: emoji 或单字头像
        tool_policy:
          $ref: "#/components/schemas/ToolPolicy"
        grants:
          type: array
          description: 仅人设 admin 可见；无 admin 时省略该键。
          items:
            $ref: "#/components/schemas/Grant"
    Grant:
      type: object
      required: [user_id, capabilities]
      properties:
        user_id:
          type: string
          format: uuid
        capabilities:
          type: array
          items:
            type: string
            enum: [chat, read_memory, write_memory, admin]
    ToolPolicy:
      type: object
      properties:
        allowed_tools:
          type: array
          items:
            type: string
        notes:
          type: string
...
    EventNode:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        happened_at:
          type: string
          format: date-time
        type:
          type: string
          enum: [milestone, promise, conflict, daily, work]
        importance:
          type: integer
        summary:
          type: string
        confidence:
          type: number
          format: float
        memory_ids:
          type: array
          items:
            type: string
            format: uuid
    EventEdge:
      type: object
      properties:
        from_id:
          type: string
          format: uuid
        to_id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [temporal, caused_by, involves_person]
    EventTree:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/EventNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/EventEdge"
...
    InboxItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
          enum: [fact, event, conflict]
        payload:
          type: object
          additionalProperties: true
        conflict_with:
          type: object
          additionalProperties: true
          nullable: true
        status:
          type: string
          enum: [pending, confirmed, dismissed]
    ImportJob:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [queued, running, completed, failed]
        error:
          type: string
          nullable: true
    MemoryItem:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
          enum: [fact, episode_summary, file_chunk, image_caption, transcript]
        text:
          type: string
        event_id:
          type: string
          format: uuid
          nullable: true
        status:
          type: string
          enum: [active, superseded, deleted]
        created_at:
          type: string
          format: date-time
  responses:
    Error:
      description: 错误
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
security:
  - bearerAuth: []
paths:
...
  /personas:
    get:
      tags: [Personas]
      summary: 列出人设
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      responses:
        "200":
          description: 人设列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Persona"
    post:
      tags: [Personas]
      summary: 创建人设
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [display_name, skin]
              properties:
                skin:
                  $ref: "#/components/schemas/PersonaSkin"
                display_name:
                  type: string
                one_liner:
                  type: string
                personality:
                  type: object
                  additionalProperties: true
                taboos:
                  type: array
                  items:
                    type: string
                relationships:
                  type: array
                  items:
                    type: object
                    additionalProperties: true
      responses:
        "201":
          description: 已创建
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
        "403":
          $ref: "#/components/responses/Error"
  /personas/{persona_id}:
    get:
      tags: [Personas]
      summary: 人设详情（无 read_memory 时为最小档案）
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      responses:
        "200":
          description: 人设
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Persona"
        "404":
          $ref: "#/components/responses/Error"
    patch:
      tags: [Personas]
      summary: 更新档案
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: 已更新
        "403":
          $ref: "#/components/responses/Error"
  /personas/{persona_id}/grants:
    put:
      tags: [Personas]
      summary: 覆盖人设授权
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [grants]
              properties:
                grants:
                  type: array
                  items:
                    $ref: "#/components/schemas/Grant"
      responses:
        "200":
          description: 已覆盖
        "403":
          $ref: "#/components/responses/Error"
...
  /personas/{persona_id}/memories:
    get:
      tags: [Memory]
      summary: 列出记忆
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
        - name: type
          in: query
          schema:
            type: string
        - name: event_id
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            default: active
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: 记忆列表
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/MemoryItem"
                  total:
                    type: integer
                    description: 过滤后、分页前的条数
        "403":
          $ref: "#/components/responses/Error"
  /personas/{persona_id}/memories/{memory_id}:
    delete:
      tags: [Memory]
      summary: 删除记忆
      description: 需要 persona 的 admin 能力；软删除记忆并移除向量索引。
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
        - name: memory_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "204":
          description: 已删除
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
  /personas/{persona_id}/imports:
    post:
      tags: [Memory]
      summary: 导入多模态资料
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: 默认上限 32MB
                hint:
                  type: string
      responses:
        "202":
          description: 已入队
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "403":
          $ref: "#/components/responses/Error"
  /imports/{job_id}:
    get:
      tags: [Memory]
      summary: 导入任务状态
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: job_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 任务
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportJob"
        "404":
          $ref: "#/components/responses/Error"
  /personas/{persona_id}/inbox:
    get:
      tags: [Memory]
      summary: 记忆确认收件箱
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      responses:
        "200":
          description: Inbox
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/InboxItem"
        "403":
          $ref: "#/components/responses/Error"
  /personas/{persona_id}/inbox/bootstrap:
    post:
      tags: [Memory]
      summary: 从待确认 Inbox 生成档案初稿与首批事件节点
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
      responses:
        "200":
          description: 处理结果
          content:
            application/json:
              schema:
                type: object
                properties:
                  profile_updated:
                    type: boolean
                  events_created:
                    type: integer
                  memories_created:
                    type: integer
                  inbox_processed:
                    type: integer
        "403":
          $ref: "#/components/responses/Error"
        "404":
          $ref: "#/components/responses/Error"
...
  /inbox/{inbox_id}/confirm:
    post:
      tags: [Memory]
      summary: 确认写入档案或事件树
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: inbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                mark_key_event:
                  type: boolean
                  default: false
      responses:
        "200":
          description: 已确认
        "409":
          $ref: "#/components/responses/Error"
  /inbox/{inbox_id}/dismiss:
    post:
      tags: [Memory]
      summary: 忽略候选记忆
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: inbox_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 已忽略
        "409":
          $ref: "#/components/responses/Error"
  /personas/{persona_id}/events/tree:
    get:
      tags: [Events]
      summary: 事件树或时间轴投影
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - $ref: "#/components/parameters/PersonaId"
        - name: view
          in: query
          schema:
            type: string
            enum: [tree, timeline]
            default: tree
        - name: key_only
          in: query
          schema:
            type: boolean
            default: true
      responses:
        "200":
          description: 树
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventTree"
        "403":
          $ref: "#/components/responses/Error"
  /events/{event_id}:
    get:
      tags: [Events]
      summary: 事件卡
      parameters:
        - $ref: "#/components/parameters/TenantHeader"
        - name: event_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: 事件卡
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
                    format: uuid
                  persona_id:
                    type: string
                    format: uuid
                  title:
                    type: string
                  summary:
                    type: string
                  confidence:
                    type: number
                  participants:
                    type: array
                    items:
                      type: string
                  causal_in:
                    type: array
                    items:
                      type: object
                  causal_out:
                    type: array
                    items:
                      type: object
                  verbatim:
                    type: array
                    items:
                      type: object
                  attachments:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        type:
                          type: string
                        text:
                          type: string
                  memories:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          format: uuid
                        text:
                          type: string
        "404":
          $ref: "#/components/responses/Error"
...